Add multiplage configuration  - /marscalendar page  - Resize and phi rotation for the mars TER image (models/cropResized.jpg)
