+++ src/index.html
@@ -12,7 +12,6 @@
     <link rel="icon" href="./img/favicon.svg" type="image/svg+xml">
     <title>ARESIAN(Mars Geography Simulator)</title>
 
-
 </head>
 <body>
     <audio id="bgmusic" ref="player" volume="0.1">
