+++ src/marscalendar/marscalendar.html
@@ -2,7 +2,7 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Moon Calendar</title>
+    <title>Mars Calendar</title>
 
     <style>
